lint
SCCS-vsn: sys/vm/vm_page.c 7.7
SCCS-vsn: sys/vm/vm_object.c 7.6

+++ usr/src/sys/vm/vm_object.c
@@ -7,7 +7,7 @@
  *
  * %sccs.include.redist.c%
  *
- *	@(#)vm_object.c	7.5 (Berkeley) %G%
+ *	@(#)vm_object.c	7.6 (Berkeley) %G%
  *
  *
  * Copyright (c) 1987, 1990 Carnegie-Mellon University.
@@ -282,7 +282,7 @@ void vm_object_terminate(object)
 	 */
 
 	while (object->paging_in_progress != 0) {
-		vm_object_sleep(object, object, FALSE);
+		vm_object_sleep((int)object, object, FALSE);
 		vm_object_lock(object);
 	}
